requirements.txt
discord
python-dotenv
Flask
asgiref
selenium
webdriver-manager